pyproject.toml
[project]
name = "wso"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "towienko", email = "tomek.owienko@gmail.com"},
]
dependencies = ["libvirt-python>=11.3.0", "python-dotenv>=1.1.0", "python-daemon>=3.1.2"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "ruff>=0.11.11",
    "mypy>=1.15.0",
]

[tool.mypy]
exclude = [".venv"]
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "libvirt"
ignore_missing_imports = true

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
target-version = "py313"
indent-width = 4
line-length = 120

[tool.ruff.lint]
# By default, Ruff enables Flake8's E and F rules
# Pyflakes - F, pycodestyle - E, W
# flake8-builtins - A
# Pylint - PLC, PLE, PLW
# isort - I
select = ['E', 'F', 'W', 'A', 'PLC', 'PLE', 'PLW', 'I', 'FAST']
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
